+++ libs/brain/questionnaire/src/lib/brn-questionnaire-choice-label.ts
@@ -1,0 +1,12 @@
+import { Directive } from '@angular/core';
+import { injectBrnQuestionnaireChoice } from './brn-questionnaire.token';
+
+@Directive({
+	selector: '[brnQuestionnaireChoiceLabel]',
+	exportAs: 'brnQuestionnaireChoiceLabel',
+})
+export class BrnQuestionnaireChoiceLabel {
+	constructor() {
+		injectBrnQuestionnaireChoice();
+	}
+}